refactor: auth 패키지를 최상위 컨텍스트로 승격하고 ddd 계층 구조 적용
