+++ src/main/java/com/techfork/auth/application/dto/KakaoLoginResponse.java
@@ -1,4 +1,4 @@
-package com.techfork.domain.auth.dto;
+package com.techfork.auth.application.dto;
 
 import lombok.Builder;
