+++ src/main/java/com/techfork/auth/application/dto/TokenRefreshResponse.java
@@ -1,4 +1,4 @@
-package com.techfork.domain.auth.dto;
+package com.techfork.auth.application.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
