+++ src/main/java/com/techfork/domain/admin/controller/AdminController.java
@@ -1,7 +1,7 @@
 package com.techfork.domain.admin.controller;
 
-import com.techfork.domain.auth.dto.DeveloperTokenResponse;
-import com.techfork.domain.auth.service.AuthService;
+import com.techfork.auth.application.dto.DeveloperTokenResponse;
+import com.techfork.auth.application.AuthService;
 import com.techfork.domain.source.service.CrawlingService;
 import com.techfork.global.common.code.SuccessCode;
 import com.techfork.global.response.BaseResponse;
