+++ src/test/java/com/techfork/domain/auth/controller/AuthControllerIntegrationTest.java
@@ -3,8 +3,8 @@ package com.techfork.domain.auth.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
-import com.techfork.domain.auth.dto.KakaoLoginRequest;
-import com.techfork.domain.auth.exception.AuthErrorCode;
+import com.techfork.auth.application.dto.KakaoLoginRequest;
+import com.techfork.auth.domain.exception.AuthErrorCode;
 import com.techfork.useraccount.domain.User;
 import com.techfork.useraccount.domain.enums.Role;
 import com.techfork.useraccount.domain.enums.SocialType;
