+++ src/test/java/com/techfork/domain/auth/service/KakaoOAuthServiceTest.java
@@ -1,7 +1,8 @@
 package com.techfork.domain.auth.service;
 
-import com.techfork.domain.auth.dto.kakao.KakaoUserInfoResponse;
-import com.techfork.domain.auth.exception.AuthErrorCode;
+import com.techfork.auth.application.KakaoOAuthService;
+import com.techfork.auth.application.dto.kakao.KakaoUserInfoResponse;
+import com.techfork.auth.domain.exception.AuthErrorCode;
 import com.techfork.global.exception.GeneralException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
